Initial draft of the conditions and dev plugin complete
 - Manager exists in basic form but needs interface work
 - plugins can declare components to be exposed by the plugin manager
 - plugins can declare where their src is - for linting purposes
 - Example plugins being brought up to main code base standards
 - fallbacks added to tools in case pluggy not found
