+++ plugins/mewbot-client_for_reddit/src/mewbot_reddit/__init__.py
@@ -2,53 +2,15 @@
 
 from typing import Tuple, Type, Dict
 
+import os
+
 from mewbot.api.v1 import IOConfig, Input, InputEvent
 from mewbot.plugins.hook_specs import mewbot_ext_hook_impl
+from mewbot.plugins.hook_specs import mewbot_dev_hook_impl
 
 from .io_configs.reddit_password_io import RedditPasswordIO
-from .io_configs import RedditSubredditInput, RedditRedditorInput
-from .input_events import (
-    # subreddit first events
-    # - a submission has been manipulated in a subreddit
-    SubRedditSubmissionCreationInputEvent,
-    SubRedditSubmissionEditInputEvent,
-    SubRedditSubmissionDeletedInputEvent,
-    SubRedditSubmissionRemovedInputEvent,
-    SubRedditSubmissionPinnedInputEvent,  # Todo: Implement this
-    # - a comment has been manipulated in the subreddit
-    SubRedditCommentCreationInputEvent,
-    SubRedditCommentEditInputEvent,
-    SubRedditCommentDeletedInputEvent,
-    SubRedditCommentRemovedInputEvent,
-    # - a user has interacted with the subreddit
-    RedditUserJoinedSubredditInputEvent,
-    RedditUserLeftSubredditInputEvent,
-    RedditUserBannedFromSubredditInputEvent,
-    # user first events
-    # - a user has manipulated a submission in a subreddit
-    RedditUserCreatedSubredditSubmissionInputEvent,
-    RedditUserEditedSubredditSubmissionInputEvent,
-    RedditUserDeletedSubredditSubmissionInputEvent,
-    RedditUserRemovedSubredditSubmissionInputEvent,
-    # - a user has manipulated a comment in a subreddit
-    RedditUserCreatedCommentOnSubredditSubmissionInputEvent,
-    RedditUserEditedCommentOnSubredditSubmissionInputEvent,
-    RedditUserDeletedCommentOnSubredditSubmissionInputEvent,
-    RedditUserRemovedCommentOnSubredditSubmissionInputEvent,
-    # - a user has manipulated a submission in their profile
-    RedditUserCreatedProfileSubmissionInputEvent,
-    RedditUserEditProfileSubmissionInputEvent,
-    RedditUserDeletedProfileSubmissionInputEvent,
-    RedditUserRemovedProfileSubmissionInputEvent,
-    # - a user has manipulated a comment in their profile
-    RedditUserCreatedCommentOnProfileSubmissionInputEvent,
-    RedditUserEditedCommentOnProfileSubmissionInputEvent,
-    RedditUserDeletedCommentOnProfileSubmissionInputEvent,
-    RedditUserRemovedCommentOnProfileSubmissionInputEvent,
-    # - Not entirely sure how to implement this
-    RedditPersonaVoteInputEvent,
-    RedditPostVoteInputEvent,
-)
+from .io_configs import RedditRedditorInput, USED_INPUT_EVENTS
+from .io_configs.inputs.subreddit import RedditSubredditInput
 
 # This is the name which will actually show up in the plugin manager.
 # Note - this also allows you to extend an existing plugin - just set the name
@@ -65,11 +27,12 @@ def get_io_config_classes() -> Dict[str, Tuple[Type[IOConfig], ...]]:
     be generated from properties.
     :return:
     """
-    return {__mewbot_plugin_name__: tuple(
-        [
-            RedditPasswordIO,
-        ]
-    )
+    return {
+        __mewbot_plugin_name__: tuple(
+            [
+                RedditPasswordIO,
+            ]
+        )
     }
 
 
@@ -83,51 +46,27 @@ def get_input_classes() -> Dict[str, Tuple[Type[Input], ...]]:
     return {__mewbot_plugin_name__: tuple([RedditSubredditInput, RedditRedditorInput])}
 
 
-@mewbot_ext_hook_impl
+@mewbot_ext_hook_impl  # type: ignore
 def get_input_event_classes() -> Dict[str, Tuple[Type[InputEvent], ...]]:
     """
     Returns all the InputEvent subclasses defined by this plugin.
     :return:
     """
-    return {__mewbot_plugin_name__: tuple([
-        # subreddit first events
-        # - a submission has been manipulated in a subreddit
-        SubRedditSubmissionCreationInputEvent,
-        SubRedditSubmissionEditInputEvent,
-        SubRedditSubmissionDeletedInputEvent,
-        SubRedditSubmissionRemovedInputEvent,
-        SubRedditSubmissionPinnedInputEvent,  # Todo: Implement this
-        # - a comment has been manipulated in the subreddit
-        SubRedditCommentCreationInputEvent,
-        SubRedditCommentEditInputEvent,
-        SubRedditCommentDeletedInputEvent,
-        SubRedditCommentRemovedInputEvent,
-        # - a user has interacted with the subreddit
-        RedditUserJoinedSubredditInputEvent,
-        RedditUserLeftSubredditInputEvent,
-        RedditUserBannedFromSubredditInputEvent,
-        # user first events
-        # - a user has manipulated a submission in a subreddit
-        RedditUserCreatedSubredditSubmissionInputEvent,
-        RedditUserEditedSubredditSubmissionInputEvent,
-        RedditUserDeletedSubredditSubmissionInputEvent,
-        RedditUserRemovedSubredditSubmissionInputEvent,
-        # - a user has manipulated a comment in a subreddit
-        RedditUserCreatedCommentOnSubredditSubmissionInputEvent,
-        RedditUserEditedCommentOnSubredditSubmissionInputEvent,
-        RedditUserDeletedCommentOnSubredditSubmissionInputEvent,
-        RedditUserRemovedCommentOnSubredditSubmissionInputEvent,
-        # - a user has manipulated a submission in their profile
-        RedditUserCreatedProfileSubmissionInputEvent,
-        RedditUserEditProfileSubmissionInputEvent,
-        RedditUserDeletedProfileSubmissionInputEvent,
-        RedditUserRemovedProfileSubmissionInputEvent,
-        # - a user has manipulated a comment in their profile
-        RedditUserCreatedCommentOnProfileSubmissionInputEvent,
-        RedditUserEditedCommentOnProfileSubmissionInputEvent,
-        RedditUserDeletedCommentOnProfileSubmissionInputEvent,
-        RedditUserRemovedCommentOnProfileSubmissionInputEvent,
-        # - Not entirely sure how to implement this
-        RedditPersonaVoteInputEvent,
-        RedditPostVoteInputEvent,
-    ])}
+    return {__mewbot_plugin_name__: USED_INPUT_EVENTS}
+
+
+@mewbot_dev_hook_impl  # type: ignore
+def declare_src_locs() -> Tuple[str, ...]:
+    """
+    If we declare the location of this plugin's source tree then it can be linted.
+    :return:
+    """
+    current_file = __file__
+    mewbot_reddit_top_level_folder = str(os.path.split(current_file)[0])
+    mewbot_reddit_src_folder = str(os.path.split(mewbot_reddit_top_level_folder)[0])
+
+    return tuple(
+        [
+            mewbot_reddit_src_folder,
+        ]
+    )
